+++ nodes/EasyRedmine/operations/GetOneOperation.ts
@@ -1,6 +1,6 @@
 import { IExecuteFunctions, IHttpRequestOptions } from 'n8n-workflow';
 import { EasyNodeResourceType } from '../Model';
-import { sanitizeDomain } from '../utils/SanitizeDomain';
+import { sanitizeDomain } from '../utils';
 
 export async function processGetOneOperation(
 	this: IExecuteFunctions,