+++ nodes/EasyRedmine/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './ConvertToEasyDate';
+export * from './GetProjectId';
 export * from './ExtractBillingOptions';
 export * from './SanitizeDomain';
 export * from './TryToParseParameterAsNumber';